Update API descriptions

+++ server/src/category/categoryService.ts
@@ -1,22 +1,23 @@
 import { Category } from '@prisma/client';
 import { prisma } from '../../prisma/prisma'; // Importing the shared Prisma instance
 import { CategoryDto } from './categoryDto';
+import { NotFoundError } from '../errors';
 
 export class CategoryService {
 
     public async getCategories(): Promise<Category[]> {
-        try {
-            return await prisma.category.findMany();
-        }
-        catch (error) {
-            throw error
-        }
+        return await prisma.category.findMany();
     }
 
     public async getCategory(categoryId: number): Promise<Category | null> {
-        return await prisma.category.findFirst({
-            where: { id: categoryId },
+        const category = await prisma.category.findUnique({
+            where: { id: categoryId }
         });
+        if (!category) {
+            throw new NotFoundError(`Category with id ${categoryId} not found`);
+        }
+        
+        return category
     }
 
     public async createCategory(category: CategoryDto): Promise<Category> {
@@ -25,14 +26,28 @@ export class CategoryService {
         });
     }
 
-    public async updateCategory(categoryId: number, category: CategoryDto): Promise<Category | null> {
+    public async updateCategory(categoryId: number, categoryData: CategoryDto): Promise<Category | null> {
+        const category = await prisma.category.findUnique({
+            where: { id: categoryId }
+        });
+        if (!category) {
+            throw new NotFoundError(`Category with id ${categoryId} not found`);
+        }
+
         return await prisma.category.update({
             where: { id: categoryId },
-            data: category,
+            data: categoryData,
         });
     }
 
     public async deleteCategory(categoryId: number): Promise<void> {
+        const category = await prisma.category.findUnique({
+            where: { id: categoryId }
+        });
+        if (!category) {
+            throw new NotFoundError(`Category with id ${categoryId} not found`);
+        }
+        
         await prisma.category.delete({
             where: { id: categoryId },
         });
